[ADD] Fucntion for automatic package creation, item quantity function adjustment and access right rearrangement

+++ ihhg_master/models/scm_entry_package_line.py
@@ -9,7 +9,7 @@ class SelectionCriterium(models.Model):
     scm_id = fields.Many2one(comodel_name="scm.entry")
     package_id = fields.Many2one('ihh.package', string='Package', domain="[('delivery_address_ids', 'in', delivery_address_id)]")
     delivery_address_id = fields.Many2one(comodel_name="res.partner", domain="[('package_id', '=', package_id)]")
-    quantity = fields.Float()
+    quantity = fields.Integer()
     scm_entry_item_line_id = fields.One2many(comodel_name="scm.entry.item.line", inverse_name="scm_package_line_id", readonly=True)
 
     # TODO see if more complexe name need to be done
